reuse middleware and client_socket function in iodine integration, and minor refactor
