+++ lib/lite_cable/connection/base.rb
@@ -75,7 +75,6 @@ module LiteCable
         command = decode(websocket_message)
         subscriptions.execute_command command
       rescue Subscriptions::Error, Channel::Error, Channel::Registry::Error => e
-        # TODO: может backtrace сюда добавить?
         log(:error, log_fmt("Connection command failed: #{e}"))
         close
       end
