+++ lib/lite_cable/server/client_socket.rb
@@ -4,6 +4,7 @@ module LiteCable
   module Server
     module ClientSocket # :nodoc:
       require "lite_cable/server/client_socket/subscriptions"
+      require "lite_cable/server/client_socket/handlers"
       require "lite_cable/server/client_socket/base"
     end
   end
