+++ lib/lite_cable/server/subscribers_map.rb
@@ -1,4 +1,3 @@
-# coding: utf-8
 # frozen_string_literal: true
 
 module LiteCable
